+++ lib/pubspec2.dart
@@ -1,0 +1,18 @@
+// Copyright (c) 2015, Anders Holmgren. All rights reserved. Use of this
+//source code
+// is governed by a BSD-style license that can be found in the LICENSE file.
+
+/// The pubspec library.
+///
+/// A library for manipulating [pubspec](https://www.dartlang.org/tools/pub/pubspec.html) files
+library;
+
+export 'src/dependency/dependency.dart';
+export 'src/dependency/external_hosted_reference.dart';
+export 'src/dependency/git_reference.dart';
+export 'src/dependency/hosted_reference.dart';
+export 'src/dependency/path_reference.dart';
+export 'src/dependency/sdk_reference.dart';
+export 'src/executable.dart';
+export 'src/pubspec.dart';
+export 'src/yaml_to_string.dart';