+++ lib/src/pubspec.dart
@@ -1,19 +1,20 @@
-// Copyright (c) 2015, Anders Holmgren. All rights reserved. Use of this source code
+// Copyright (c) 2015, Anders Holmgren. All rights reserved.
+// Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-library pubspec.spec;
-
 import 'dart:async';
-import 'dart:io';
+import 'dart:io' hide Platform;
 
 import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
 import 'package:yaml/yaml.dart';
 
+import 'dependency/dependency.dart';
 import 'executable.dart';
 import 'json_utils.dart';
-import 'dependency.dart';
+import 'platform.dart';
 import 'yaml_to_string.dart';
+import 'yaml_util.dart';
 
 /// Represents a [pubspec](https://www.dartlang.org/tools/pub/pubspec.html).
 ///
@@ -31,6 +32,45 @@ import 'yaml_to_string.dart';
 ///
 ///
 class PubSpec implements Jsonable {
+  const PubSpec(
+      {this.name,
+      this.author,
+      this.version,
+      this.homepage,
+      this.documentation,
+      this.description,
+      this.publishTo,
+      this.environment,
+      this.dependencies = const {},
+      this.devDependencies = const {},
+      this.dependencyOverrides = const {},
+      this.executables = const {},
+      this.platforms = const {},
+      this.unParsedYaml = const {}});
+
+  factory PubSpec.fromYaml(YamlMap yamlMap) {
+    final p = parseYaml(yamlMap, consumeMap: true);
+
+    return PubSpec(
+        name: p.single('name') as String?,
+        author: p.single('author') as String?,
+        version: p.single('version', Version.parse),
+        homepage: p.single('homepage') as String?,
+        documentation: p.single('documentation') as String?,
+        description: p.single('description') as String?,
+        publishTo: p.single('publish_to', Uri.parse),
+        environment: p.single('environment', Environment.fromJson),
+        dependencies: p.getReferences('dependencies'),
+        devDependencies: p.getReferences('dev_dependencies'),
+        dependencyOverrides: p.getReferences('dependency_overrides'),
+        executables: p.getPairs<Executable>('executables', Executable.fromJson),
+        platforms:
+            p.getPairs<Platform>('platforms', (k, v) => Platform.fromJson(k)),
+        unParsedYaml: p.unconsumed);
+  }
+
+  factory PubSpec.fromYamlString(String yamlString) =>
+      PubSpec.fromYaml(loadYaml(yamlString) as YamlMap);
   final String? name;
 
   final String? author;
@@ -71,54 +111,17 @@ class PubSpec implements Jsonable {
 
   final Map<String, Executable> executables;
 
-  final Map? unParsedYaml;
+  final Map<String, Platform> platforms;
 
-  const PubSpec(
-      {this.name,
-      this.author,
-      this.version,
-      this.homepage,
-      this.documentation,
-      this.description,
-      this.publishTo,
-      this.environment,
-      this.dependencies: const {},
-      this.devDependencies: const {},
-      this.dependencyOverrides: const {},
-      this.executables: const {},
-      this.unParsedYaml: const {}});
-
-  factory PubSpec.fromJson(Map? json) {
-    final p = parseJson(json, consumeMap: true);
-    return PubSpec(
-        name: p.single('name'),
-        author: p.single('author'),
-        version: p.single('version', (v) => Version.parse(v)),
-        homepage: p.single('homepage'),
-        documentation: p.single('documentation'),
-        description: p.single('description'),
-        publishTo: p.single('publish_to', (v) => Uri.parse(v)),
-        environment: p.single('environment', (v) => Environment.fromJson(v)),
-        dependencies:
-            p.mapValues('dependencies', (v) => DependencyReference.fromJson(v)),
-        devDependencies: p.mapValues(
-            'dev_dependencies', (v) => DependencyReference.fromJson(v)),
-        dependencyOverrides: p.mapValues(
-            'dependency_overrides', (v) => DependencyReference.fromJson(v)),
-        executables: p.mapValues('executables', (v) => Executable.fromJson(v)),
-        unParsedYaml: p.unconsumed);
-  }
-
-  factory PubSpec.fromYamlString(String yamlString) =>
-      PubSpec.fromJson(loadYaml(yamlString));
+  final Json? unParsedYaml;
 
   /// loads the pubspec from the [projectDirectory]
-  static Future<PubSpec> load(Directory projectDirectory) =>
-      loadFile(p.join(projectDirectory.path, 'pubspec.yaml'));
+  static Future<PubSpec> load(String projectDirectory) =>
+      loadFile(p.join(projectDirectory, 'pubspec.yaml'));
 
   /// loads the pubspec from the [file]
   static Future<PubSpec> loadFile(String file) async =>
-      PubSpec.fromJson(loadYaml(await File(file).readAsString()));
+      PubSpec.fromYaml(loadYaml(await File(file).readAsString()) as YamlMap);
 
   /// creates a copy of the pubspec with the changes provided
   PubSpec copy({
@@ -134,73 +137,69 @@ class PubSpec implements Jsonable {
     Map<String, DependencyReference>? devDependencies,
     Map<String, DependencyReference>? dependencyOverrides,
     Map<String, Executable>? executables,
-    Map? unParsedYaml,
-  }) {
-    return PubSpec(
-        name: name ?? this.name,
-        author: author ?? this.author,
-        version: version ?? this.version,
-        homepage: homepage ?? this.homepage,
-        documentation: documentation ?? this.documentation,
-        description: description ?? this.description,
-        publishTo: publishTo ?? this.publishTo,
-        environment: environment ?? this.environment,
-        dependencies: dependencies ?? this.dependencies,
-        devDependencies: devDependencies ?? this.devDependencies,
-        dependencyOverrides: dependencyOverrides ?? this.dependencyOverrides,
-        executables: executables ?? this.executables,
-        unParsedYaml: unParsedYaml ?? this.unParsedYaml);
-  }
+    Map<String, Platform>? platforms,
+    Json? unParsedYaml,
+  }) =>
+      PubSpec(
+          name: name ?? this.name,
+          author: author ?? this.author,
+          version: version ?? this.version,
+          homepage: homepage ?? this.homepage,
+          documentation: documentation ?? this.documentation,
+          description: description ?? this.description,
+          publishTo: publishTo ?? this.publishTo,
+          environment: environment ?? this.environment,
+          dependencies: dependencies ?? this.dependencies,
+          devDependencies: devDependencies ?? this.devDependencies,
+          dependencyOverrides: dependencyOverrides ?? this.dependencyOverrides,
+          executables: executables ?? this.executables,
+          platforms: platforms ?? this.platforms,
+          unParsedYaml: unParsedYaml ?? this.unParsedYaml);
 
   /// saves the pubspec to the [projectDirectory]
-  Future save(Directory projectDirectory) async {
-    final ioSink =
-        File(p.join(projectDirectory.path, 'pubspec.yaml')).openWrite();
+  Future<dynamic> save(String projectDirectory) async {
+    final ioSink = File(p.join(projectDirectory, 'pubspec.yaml')).openWrite();
     try {
-      YamlToString().writeYamlString(toJson(), ioSink);
+      const YamlToString().writeYamlString(toJson(), ioSink);
     } finally {
-      return ioSink.close();
+      await ioSink.flush();
+      await ioSink.close();
     }
   }
 
   /// Converts to a Map that can be serialised to Yaml or Json
   @override
-  Map toJson() {
-    return (buildJson
-          ..add('name', name)
-          ..add('author', author)
-          ..add('version', version)
-          ..add('homepage', homepage)
-          ..add('documentation', documentation)
-          ..add('publish_to', publishTo)
-          ..add('environment', environment)
-          ..add('description', description)
-          ..add('dependencies', dependencies)
-          ..add('dev_dependencies', devDependencies)
-          ..add('dependency_overrides', dependencyOverrides)
-          ..add('executables', executables)
-          ..addAll(unParsedYaml!))
-        .json;
-  }
+  Json toJson() => (buildJson
+        ..add('name', name)
+        ..add('author', author)
+        ..add('version', version)
+        ..add('homepage', homepage)
+        ..add('documentation', documentation)
+        ..add('publish_to', publishTo)
+        ..add('environment', environment)
+        ..add('description', description)
+        ..add('dependencies', dependencies)
+        ..add('dev_dependencies', devDependencies)
+        ..add('dependency_overrides', dependencyOverrides)
+        ..add('executables', executables)
+        ..add('platforms', platforms)
+        ..addAll(unParsedYaml!))
+      .json;
 }
 
 class Environment implements Jsonable {
-  final VersionConstraint? sdkConstraint;
-  final Map? unParsedYaml;
-
   const Environment(this.sdkConstraint, this.unParsedYaml);
 
-  factory Environment.fromJson(Map json) {
+  factory Environment.fromJson(Json json) {
     final p = parseJson(json, consumeMap: true);
-    return Environment(
-        p.single('sdk', (v) => VersionConstraint.parse(v)), p.unconsumed);
+    return Environment(p.single('sdk', VersionConstraint.parse), p.unconsumed);
   }
+  final VersionConstraint? sdkConstraint;
+  final Json? unParsedYaml;
 
   @override
-  Map toJson() {
-    return (buildJson
-          ..add('sdk', "${sdkConstraint.toString()}")
-          ..addAll(unParsedYaml!))
-        .json;
-  }
+  Json toJson() => (buildJson
+        ..add('sdk', '$sdkConstraint')
+        ..addAll(unParsedYaml!))
+      .json;
 }