+++ lib/src/yaml_to_string.dart
@@ -1,3 +1,5 @@
+import 'dependency/dependency.dart';
+
 final _unsuportedCharacters = RegExp(
     r'''^[\n\t ,[\]{}#&*!|<>'"%@']|^[?-]$|^[?-][ \t]|[\n:][ \t]|[ \t]\n|[\n\t ]#|[\n\t :]$''');
 
@@ -7,34 +9,35 @@ class YamlToString {
     this.quotes = "'",
   });
 
-  final String indent, quotes;
-  static final _divider = ': ';
+  final String indent;
+  final String quotes;
+  static const _divider = ': ';
 
-  String toYamlString(node) {
+  String toYamlString(Object node) {
     final stringBuffer = StringBuffer();
     writeYamlString(node, stringBuffer);
     return stringBuffer.toString();
   }
 
   /// Serializes [node] into a String and writes it to the [sink].
-  void writeYamlString(node, StringSink sink) {
+  void writeYamlString(Object node, StringSink sink) {
     _writeYamlString(node, 0, sink, true);
   }
 
   void _writeYamlString(
-    node,
+    Object node,
     int indentCount,
     StringSink stringSink,
     bool isTopLevel,
   ) {
-    if (node is Map) {
+    if (node is Json) {
       _mapToYamlString(node, indentCount, stringSink, isTopLevel);
-    } else if (node is Iterable) {
+    } else if (node is Iterable<String>) {
       _listToYamlString(node, indentCount, stringSink, isTopLevel);
     } else if (node is String) {
       stringSink.writeln(_escapeString(node));
     } else if (node is double) {
-      stringSink.writeln("!!float $node");
+      stringSink.writeln('!!float $node');
     } else {
       stringSink.writeln(node);
     }
@@ -51,7 +54,7 @@ class YamlToString {
   }
 
   void _mapToYamlString(
-    node,
+    Json node,
     int indentCount,
     StringSink stringSink,
     bool isTopLevel,
@@ -63,19 +66,21 @@ class YamlToString {
 
     final keys = _sortKeys(node);
 
-    keys.forEach((key) {
+    for (final key in keys) {
       final value = node[key];
       _writeIndent(indentCount, stringSink);
-      stringSink..write(key)..write(_divider);
-      _writeYamlString(value, indentCount, stringSink, false);
-    });
+      stringSink
+        ..write(key)
+        ..write(_divider);
+      _writeYamlString(value ?? '', indentCount, stringSink, false);
+    }
   }
 
-  Iterable<String> _sortKeys(Map map) {
-    final simple = <String>[],
-        maps = <String>[],
-        lists = <String>[],
-        other = <String>[];
+  Iterable<String> _sortKeys(Json map) {
+    final simple = <String>[];
+    final maps = <String>[];
+    final lists = <String>[];
+    final other = <String>[];
 
     map.forEach((key, value) {
       if (value is String) {
@@ -93,7 +98,7 @@ class YamlToString {
   }
 
   void _listToYamlString(
-    Iterable node,
+    Iterable<Object> node,
     int indentCount,
     StringSink stringSink,
     bool isTopLevel,
@@ -103,11 +108,11 @@ class YamlToString {
       indentCount += 2;
     }
 
-    node.forEach((value) {
+    for (final value in node) {
       _writeIndent(indentCount, stringSink);
       stringSink.write('- ');
       _writeYamlString(value, indentCount, stringSink, false);
-    });
+    }
   }
 
   void _writeIndent(int indentCount, StringSink stringSink) =>