+++ test/dependency_test.dart
@@ -1,13 +1,12 @@
-import 'dart:io';
-
+import 'package:dcli_core/dcli_core.dart';
 import 'package:pub_semver/pub_semver.dart';
-import 'package:pubspec/pubspec.dart';
+import 'package:pubspec2/pubspec2.dart';
 import 'package:test/test.dart';
 
-main() {
+void main() {
   group('external hosted dependency', () {
     test('fromYamlString ( sdk < 2.15 )', () {
-      var pubspecString = 'name: my_test_lib\n'
+      const pubspecString = 'name: my_test_lib\n'
           'version: 0.1.0\n'
           'description: for testing\n'
           'dependencies:\n'
@@ -17,18 +16,18 @@ main() {
           '            name: custom_lib\n'
           '            url: https://pub.mycompany.org\n'
           '        version: ^0.1.0';
-      var p = PubSpec.fromYamlString(pubspecString);
-      var dep = p.dependencies['self_hosted_lib']!;
-      expect(dep, TypeMatcher<ExternalHostedReference>());
+      final p = PubSpec.fromYamlString(pubspecString);
+      final dep = p.dependencies['self_hosted_lib']!;
+      expect(dep, const TypeMatcher<ExternalHostedReference>());
 
-      var exDep = dep as ExternalHostedReference;
+      final exDep = dep as ExternalHostedReference;
       expect(exDep.name, 'custom_lib');
       expect(exDep.url, 'https://pub.mycompany.org');
       expect(exDep.versionConstraint.toString(), '^0.1.0');
     });
 
     test('fromYamlString ( sdk >= 2.15 )', () {
-      var pubspecString = 'name: my_test_lib\n'
+      const pubspecString = 'name: my_test_lib\n'
           'version: 0.1.0\n'
           'description: for testing\n'
           'dependencies:\n'
@@ -36,32 +35,30 @@ main() {
           '    custom_lib:\n'
           '        hosted: https://pub.mycompany.org\n'
           '        version: ^0.1.0';
-      var p = PubSpec.fromYamlString(pubspecString);
-      var dep = p.dependencies['custom_lib']!;
-      expect(dep, TypeMatcher<ExternalHostedReference>());
+      final p = PubSpec.fromYamlString(pubspecString);
+      final dep = p.dependencies['custom_lib']!;
+      expect(dep, const TypeMatcher<ExternalHostedReference>());
 
-      var exDep = dep as ExternalHostedReference;
+      final exDep = dep as ExternalHostedReference;
       expect(exDep.url, 'https://pub.mycompany.org');
       expect(exDep.versionConstraint.toString(), '^0.1.0');
     });
 
     test('to json ( sdk >= 2.15 )', () {
-      var exDep = ExternalHostedReference(
-          'custom_lib',
-          'https://pub.mycompany.org',
-          VersionConstraint.parse('^0.1.0'),
-          false);
-      var json = exDep.toJson();
+      final exDep = ExternalHostedReference('custom_lib',
+          'https://pub.mycompany.org', VersionConstraint.parse('^0.1.0'),
+          verboseFormat: false);
+      final json = exDep.toJson();
       expect(json['hosted'], 'https://pub.mycompany.org');
       expect(json['version'], '^0.1.0');
     });
 
     test('to json ( sdk < 2.15 )', () {
-      var exDep = ExternalHostedReference('custom_lib',
+      final exDep = ExternalHostedReference('custom_lib',
           'https://pub.mycompany.org', VersionConstraint.parse('^0.1.0'));
-      var json = exDep.toJson();
-      expect(json['hosted']['url'], 'https://pub.mycompany.org');
-      expect(json['hosted']['name'], 'custom_lib');
+      final json = exDep.toJson();
+      expect((json['hosted']! as Json)['url'], 'https://pub.mycompany.org');
+      expect((json['hosted']! as Json)['name'], 'custom_lib');
       expect(json['version'], '^0.1.0');
     });
   });
@@ -71,43 +68,62 @@ main() {
   /// The string any allows any version. This is equivalent to an empty
   /// version constraint, but is more explicit.
   test('dependency without the version constraint is "any" version', () {
-    var pubspecString = 'name: my_test_lib\n'
+    const pubspecString = 'name: my_test_lib\n'
         'version: 0.1.0\n'
         'description: for testing\n'
         'dependencies:\n'
         '    meta:\n';
-    var p = PubSpec.fromYamlString(pubspecString);
-    var dep = p.dependencies['meta']!;
-    expect(dep, TypeMatcher<HostedReference>());
+    final p = PubSpec.fromYamlString(pubspecString);
+    final dep = p.dependencies['meta']!;
+    expect(dep, const TypeMatcher<HostedReference>());
 
-    var exDep = dep as HostedReference;
+    final exDep = dep as HostedReference;
     expect(exDep.versionConstraint.toString(), 'any');
   });
 
   test('sdk dependency', () {
-    var pubspecString = 'name: my_test_lib\n'
+    const pubspecString = 'name: my_test_lib\n'
         'version: 0.1.0\n'
         'description: for testing\n'
         'dependencies:\n'
         '    flutter:\n'
         '        sdk: flutter\n';
-    var p = PubSpec.fromYamlString(pubspecString);
-    var dep = p.dependencies['flutter']!;
-    expect(dep, TypeMatcher<SdkReference>());
+    final p = PubSpec.fromYamlString(pubspecString);
+    final dep = p.dependencies['flutter']!;
+    expect(dep, const TypeMatcher<SdkReference>());
 
-    var sdkDep = dep as SdkReference;
+    final sdkDep = dep as SdkReference;
     expect(sdkDep.sdk, 'flutter');
   });
 
   test('load from file', () async {
-    final fromDir = await PubSpec.load(Directory('.'));
+    final fromDir = await PubSpec.load('.');
+    final fromFile = await PubSpec.loadFile('./pubspec.yaml');
+    expect(fromFile.toJson(), equals(fromDir.toJson()));
+  });
+
+  test('save to file', () async {
+    const pubspecString = 'name: my_test_lib\n'
+        'version: 0.1.0\n'
+        'description: for testing\n'
+        'dependencies:\n'
+        '    flutter:\n'
+        '        sdk: flutter\n';
+    final p = PubSpec.fromYamlString(pubspecString);
+
+    await withTempDir((tempDir) async {
+      await p.save(tempDir);
+      await PubSpec.load(tempDir);
+    });
+
+    final fromDir = await PubSpec.load('.');
     final fromFile = await PubSpec.loadFile('./pubspec.yaml');
     expect(fromFile.toJson(), equals(fromDir.toJson()));
   });
 
   group('git dependency', () {
     test('fromYamlString', () {
-      final pubspecString = 'name: my_test_lib\n'
+      const pubspecString = 'name: my_test_lib\n'
           'version: 0.1.0\n'
           'description: for testing\n'
           'dependencies:\n'
@@ -119,44 +135,65 @@ main() {
           '            path: packages/batz';
       final pubspec = PubSpec.fromYamlString(pubspecString);
 
-      var dep = pubspec.dependencies['git_lib']!;
-      expect(dep, TypeMatcher<GitReference>());
+      final dep = pubspec.dependencies['git_lib']!;
+      expect(dep, const TypeMatcher<GitReference>());
 
-      var gitDep = dep as GitReference;
+      final gitDep = dep as GitReference;
       expect(gitDep.url, 'git://github.com/foo/bar.git');
       expect(gitDep.ref, 'master');
       expect(gitDep.path, 'packages/batz');
     });
 
     test('toJson url', () {
-      final subject = GitReference('git://github.com/foo/bar.git');
+      const subject = GitReference('git://github.com/foo/bar.git');
 
-      var jsonObj = subject.toJson();
+      final jsonObj = subject.toJson();
 
       expect(jsonObj['git'], 'git://github.com/foo/bar.git');
     });
 
     test('toJson url, ref', () {
-      final subject = GitReference('git://github.com/foo/bar.git', 'master');
+      const subject = GitReference('git://github.com/foo/bar.git', 'master');
 
-      var jsonObj = subject.toJson();
+      final jsonObj = subject.toJson();
 
-      expect(jsonObj['git']['url'], 'git://github.com/foo/bar.git');
-      expect(jsonObj['git']['ref'], 'master');
+      expect((jsonObj['git']! as Json)['url'], 'git://github.com/foo/bar.git');
+      expect((jsonObj['git']! as Json)['ref'], 'master');
     });
 
     test('toJson url, ref, path', () {
-      final subject = GitReference(
+      const subject = GitReference(
         'git://github.com/foo/bar.git',
         'master',
         'packages/batz',
       );
 
-      var jsonObj = subject.toJson();
+      final jsonObj = subject.toJson();
 
-      expect(jsonObj['git']['url'], 'git://github.com/foo/bar.git');
-      expect(jsonObj['git']['ref'], 'master');
-      expect(jsonObj['git']['path'], 'packages/batz');
+      expect((jsonObj['git']! as Json)['url'], 'git://github.com/foo/bar.git');
+      expect((jsonObj['git']! as Json)['ref'], 'master');
+      expect((jsonObj['git']! as Json)['path'], 'packages/batz');
     });
+  });
+
+  test('fromYamlString ( odd order )', () {
+    const pubspecString = 'name: my_test_lib\n'
+        'version: 0.1.0\n'
+        'description: for testing\n'
+        'dependencies:\n'
+        '    meta: ^1.0.0\n'
+        '    self_hosted_lib:\n'
+        '        version: ^0.1.0\n'
+        '        hosted:\n'
+        '            name: custom_lib\n'
+        '            url: https://pub.mycompany.org\n';
+    final p = PubSpec.fromYamlString(pubspecString);
+    final dep = p.dependencies['self_hosted_lib']!;
+    expect(dep, const TypeMatcher<ExternalHostedReference>());
+
+    final exDep = dep as ExternalHostedReference;
+    expect(exDep.name, 'custom_lib');
+    expect(exDep.url, 'https://pub.mycompany.org');
+    expect(exDep.versionConstraint.toString(), '^0.1.0');
   });
 }